Add 'Next Run' menu item and extract common code for menu items to base class
