+++ src/main/kotlin/dev/arunvelsriram/desccron/DescribeCronAction.kt
@@ -1,27 +1,14 @@
 package dev.arunvelsriram.desccron
 
 import com.intellij.codeInsight.hint.HintManager
-import com.intellij.openapi.actionSystem.AnAction
 import com.intellij.openapi.actionSystem.AnActionEvent
 import com.intellij.openapi.actionSystem.PlatformDataKeys
 import com.intellij.openapi.components.service
-import com.intellij.openapi.editor.Editor
 
 
-class DescribeCronAction : AnAction() {
+class DescribeCronAction : MenuItemAction() {
     private val cronDescriptor = service<CronDescriptor>()
     private val hintManager = service<HintManager>()
-
-    override fun update(e: AnActionEvent) {
-        val editor = e.getData(PlatformDataKeys.EDITOR)
-        if (getSelectedText(editor).isEmpty()) {
-            e.presentation.isEnabled = false
-            e.presentation.isVisible = false
-        } else {
-            e.presentation.isVisible = true
-            e.presentation.isEnabled = true
-        }
-    }
 
     override fun actionPerformed(e: AnActionEvent) {
         val editor = e.getData(PlatformDataKeys.EDITOR)
@@ -32,6 +19,4 @@ class DescribeCronAction : AnAction() {
             hintManager.showErrorHint(editor!!, e.message ?: "Failed to describe cron")
         }
     }
-
-    private fun getSelectedText(editor: Editor?) = editor?.selectionModel?.selectedText?.trim() ?: ""
 }
