+++ src/main/kotlin/dev/arunvelsriram/desccron/NextRunAction.kt
@@ -1,0 +1,8 @@
+package dev.arunvelsriram.desccron;
+
+import com.intellij.openapi.actionSystem.AnActionEvent
+
+class NextRunAction : MenuItemAction() {
+    override fun actionPerformed(e: AnActionEvent) {
+    }
+}
