+++ src/test/kotlin/dev/arunvelsriram/desccron/DescribeCronActionTest.kt
@@ -40,80 +40,6 @@ internal class DescribeCronActionTest {
     }
 
     @Test
-    fun `should disable and hide when ide data is null`() {
-        every { actionEvent.getData(PlatformDataKeys.EDITOR) } returns null
-        every { actionEvent.presentation.isEnabled = false } just runs
-        every { actionEvent.presentation.isVisible = false } just runs
-        val action = DescribeCronAction()
-
-        action.update(actionEvent)
-
-        verify { actionEvent.presentation.isEnabled = false }
-        verify { actionEvent.presentation.isVisible = false }
-    }
-
-
-    @Test
-    fun `should disable and hide when selection is null`() {
-        every { actionEvent.getData(PlatformDataKeys.EDITOR) } returns editor
-        every { editor.selectionModel } returns selectionModel
-        every { selectionModel.selectedText } returns null
-        every { actionEvent.presentation.isEnabled = false } just runs
-        every { actionEvent.presentation.isVisible = false } just runs
-        val action = DescribeCronAction()
-
-        action.update(actionEvent)
-
-        verify { actionEvent.presentation.isEnabled = false }
-        verify { actionEvent.presentation.isVisible = false }
-    }
-
-    @Test
-    fun `should disable and hide when selection is empty`() {
-        every { actionEvent.getData(PlatformDataKeys.EDITOR) } returns editor
-        every { editor.selectionModel } returns selectionModel
-        every { selectionModel.selectedText } returns ""
-        every { actionEvent.presentation.isEnabled = false } just runs
-        every { actionEvent.presentation.isVisible = false } just runs
-        val action = DescribeCronAction()
-
-        action.update(actionEvent)
-
-        verify { actionEvent.presentation.isEnabled = false }
-        verify { actionEvent.presentation.isVisible = false }
-    }
-
-    @Test
-    fun `should disable and hide when selection is white spaces`() {
-        every { actionEvent.getData(PlatformDataKeys.EDITOR) } returns editor
-        every { editor.selectionModel } returns selectionModel
-        every { selectionModel.selectedText } returns "     "
-        every { actionEvent.presentation.isEnabled = false } just runs
-        every { actionEvent.presentation.isVisible = false } just runs
-        val action = DescribeCronAction()
-
-        action.update(actionEvent)
-
-        verify { actionEvent.presentation.isEnabled = false }
-        verify { actionEvent.presentation.isVisible = false }
-    }
-
-    @Test
-    fun `should enable and show when selection is not empty`() {
-        every { actionEvent.getData(PlatformDataKeys.EDITOR) } returns editor
-        every { editor.selectionModel } returns selectionModel
-        every { selectionModel.selectedText } returns "cron"
-        every { actionEvent.presentation.isEnabled = true } just runs
-        every { actionEvent.presentation.isVisible = true } just runs
-        val action = DescribeCronAction()
-
-        action.update(actionEvent)
-
-        verify { actionEvent.presentation.isEnabled = true }
-        verify { actionEvent.presentation.isVisible = true }
-    }
-
-    @Test
     fun `should show cron description as info hint`() {
         every { actionEvent.getData(PlatformDataKeys.EDITOR) } returns editor
         every { editor.selectionModel.selectedText } returns "* * * * *"
