+++ src/test/kotlin/dev/arunvelsriram/desccron/NextRunActionTest.kt
@@ -1,0 +1,3 @@
+package dev.arunvelsriram.desccron
+
+internal class NextRunActionTest
